+++ library/src/main/java/org/mustangproject/TradeParty.java
@@ -1,8 +1,8 @@
 package org.mustangproject;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -121,11 +121,11 @@ public class TradeParty implements IZUGFeRDExportableTradeParty {
 								}
 								if (party.item(partyIndex).getLocalName().equals("EndpointID")) {
 									Node currentNode = party.item(partyIndex);
-									if ((currentNode.getAttributes() != null &&
-										(currentNode.getAttributes().getNamedItem("schemeID") != null))
-										&& (party.item(partyIndex).getAttributes().getNamedItem("schemeID").getNodeValue().equals("EM"))
-									) {
-										setEmail(currentNode.getTextContent());
+									List<String> supportedSchemeIDs = Arrays.asList("EM", "0225");
+									if ((currentNode.getAttributes() != null
+											&& (currentNode.getAttributes().getNamedItem("schemeID") != null))
+											&& (supportedSchemeIDs.contains(currentNode.getAttributes().getNamedItem("schemeID").getNodeValue()))) {
+										setEmail(currentNode.getAttributes().getNamedItem("schemeID").getNodeValue(), currentNode.getTextContent());
 									}
 
 								}
@@ -163,7 +163,7 @@ public class TradeParty implements IZUGFeRDExportableTradeParty {
 												NodeList taxSchemechilds = partyTaxScheme.item(partyTaxSchemeIndex).getChildNodes();
 												for (int taxSchemechildsIndex = 0; taxSchemechildsIndex < taxSchemechilds.getLength(); taxSchemechildsIndex++) {
 													if (taxSchemechilds.item(taxSchemechildsIndex).getLocalName() != null) {
-														Set<String> taxSchemeTypes = Set.of("FC", "NOVAT");
+														List<String> taxSchemeTypes = Arrays.asList("FC", "NOVAT");
 														String textContent = taxSchemechilds.item(taxSchemechildsIndex).getTextContent();
 														if (textContent != null && taxSchemeTypes.contains(textContent)) {
 															setTaxID(CompanyId);
@@ -425,14 +425,13 @@ public class TradeParty implements IZUGFeRDExportableTradeParty {
 						}
 						if (itemChilds.item(itemChildIndex).getLocalName().equals("URIUniversalCommunication")) {
 							NodeList URIchilds = itemChilds.item(itemChildIndex).getChildNodes();
+							List<String> supportedSchemeIDs = Arrays.asList("EM", "0225");
 							for (int URIChildIndex = 0; URIChildIndex < URIchilds.getLength(); URIChildIndex++) {
 								Node currentNode = URIchilds.item(URIChildIndex);
 								if ((currentNode.getLocalName() != null) && (currentNode.getLocalName().equals("URIID")
-									&&
-									(currentNode.getAttributes().getNamedItem("schemeID") != null))
-									&& (URIchilds.item(URIChildIndex).getAttributes().getNamedItem("schemeID").getNodeValue().equals("EM"))
-								) {
-									setEmail(currentNode.getTextContent());
+										&& (currentNode.getAttributes().getNamedItem("schemeID") != null))
+										&& (supportedSchemeIDs.contains(currentNode.getAttributes().getNamedItem("schemeID").getNodeValue()))) {
+									setEmail(currentNode.getAttributes().getNamedItem("schemeID").getNodeValue(), currentNode.getTextContent());
 								}
 							}
 						}
@@ -516,6 +515,12 @@ public class TradeParty implements IZUGFeRDExportableTradeParty {
 	 */
 	public TradeParty setEmail(String eMail) {
 		SchemedID theSchemedID = new SchemedID("EM", eMail);
+		addUriUniversalCommunicationID(theSchemedID);
+		return this;
+	}
+
+	public TradeParty setEmail(String schemeID, String eMail) {
+		SchemedID theSchemedID = new SchemedID(schemeID, eMail);
 		addUriUniversalCommunicationID(theSchemedID);
 		return this;
 	}