+++ library/src/main/java/org/mustangproject/ZUGFeRD/IZUGFeRDTradeSettlementPayment.java
@@ -19,6 +19,10 @@
 package org.mustangproject.ZUGFeRD;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+
+import java.util.Arrays;
+import java.util.List;
+
 import org.mustangproject.XMLTools;
 
 public interface IZUGFeRDTradeSettlementPayment extends IZUGFeRDTradeSettlement {
@@ -62,41 +66,42 @@ public interface IZUGFeRDTradeSettlementPayment extends IZUGFeRDTradeSettlement 
 	/***
 	 * @return payment means code (BT-81 / UNTDID 4461)
 	 */
-	default String getPaymentMeansCode() { return null; };
+	default String getPaymentMeansCode() { return null; }
 
 	/***
 	 * @return payment means description (BT-82) (optional)
 	 */
-	default String getPaymentMeansInformation() { return null; };
+	default String getPaymentMeansInformation() { return null; }
 
 
 	@Override
 	@JsonIgnore
-  default String getSettlementXML() {
+	default String getSettlementXML(Profile profile) {
+		List<Profile> basicProfiles = Arrays.asList(Profiles.getByName("Basic", 1), Profiles.getByName("BasicWL", 2), Profiles.getByName("Basic", 2));
 		String accountNameStr="";
-		if (getAccountName()!=null) {
+		if (getAccountName() != null) {
 			accountNameStr="<ram:AccountName>" + XMLTools.encodeXML(getAccountName()) + "</ram:AccountName>";
-
 		}
 
 		String xml = "<ram:SpecifiedTradeSettlementPaymentMeans>"
-				+ "<ram:TypeCode>" + XMLTools.encodeXML(getPaymentMeansCode()) + "</ram:TypeCode>"
-				+ "<ram:Information>" + XMLTools.encodeXML(getPaymentMeansInformation()) + "</ram:Information>";
+				+ "<ram:TypeCode>" + XMLTools.encodeXML(getPaymentMeansCode()) + "</ram:TypeCode>";
+		if (!basicProfiles.contains(profile) && getOwnBIC() != null) {
+			xml += "<ram:Information>" + XMLTools.encodeXML(getPaymentMeansInformation()) + "</ram:Information>";
+		}
 		if (getOwnIBAN() != null) {
 			xml += "<ram:PayeePartyCreditorFinancialAccount>"
 				+ "<ram:IBANID>" + XMLTools.encodeXML(getOwnIBAN()) + "</ram:IBANID>"
 				+ accountNameStr
 				+ "</ram:PayeePartyCreditorFinancialAccount>";
 		}
-		if (getOwnBIC()!=null) {
-			xml+= "<ram:PayeeSpecifiedCreditorFinancialInstitution>"
+		if (!basicProfiles.contains(profile) && getOwnBIC() != null) {
+			xml += "<ram:PayeeSpecifiedCreditorFinancialInstitution>"
 					+ "<ram:BICID>" + XMLTools.encodeXML(getOwnBIC()) + "</ram:BICID>"
 					// + " <ram:Name>"+trans.getOwnBankName()+"</ram:Name>"
 					//
 					+ "</ram:PayeeSpecifiedCreditorFinancialInstitution>";
-
 		}
-		xml+= "</ram:SpecifiedTradeSettlementPaymentMeans>";
+		xml += "</ram:SpecifiedTradeSettlementPaymentMeans>";
 		return xml;
 	}